Add Arduino IDE sketch and upload script
Basic sketch that connects to a given WiFi network and synchronizes the rtc every hour via an NTP
server. The FQBN, port, WiFi credentials, and NTP server address are all configured in the upload.sh
shell script. This script compiles and uploads the sketch with the arduino-cli tool (part of the
Arduino IDE).