Add 2 cloud variables, fix cloud connection issues
Adds variables for setting the pitch of the low/high temperature buzzer warnings.
Calling ArduinoCloud.begin connects to the WiFi network on it's own, so there is no need to
connect ourselves. The ArduinoCloud.begin function connects asynchronously, so we wait until
connected before we connect to the NTP server. The ArduinoCloud.update function hangs if we continue
after ArduinoCloud.connected returns true, so we wait a further 10 seconds, which seems to fix the
issue.